Fix error handling in impermissible test
On failure the test throws an exception because we're missing an import
of the TestError class. Fix this by adding an import.

Signed-off-by: Bjoern Doebel <doebel@amazon.de>

Fix error handling in impermissible test

On failure the test throws an exception because we're missing an import
of the TestError class. Fix this by adding an import.

Signed-off-by: Bjoern Doebel <doebel@amazon.de>